Update i2cdisplay.cpp
Update to the default screen layout.  This makes better use of the space and breaks up the DP/LS/RS and SOCD modes so they are easier to read.  It also adds a one character space at the end of the SOCD modes allowing for us to merge the PR from DeathCore and add in multiple additional SOCD modes that we can number from 01 - XX.

Previous:

XINPUT    T15     DPAD-N

New:

XINPUT  T15  DP  SOCD-N