Add on board LED mode feature for unit status info
Feature described as:

Web-UI element that has the following options:
1 - Off
2 - Mode indicator
3 - Input test mode

Off - LED is off
Mode indicator -
LED is solid if unit is powered with connected data.
LED blinks rapidly if powered with no data.
LED blinks slowly when in web-config mode.
Input test mode -
LED is off but turns on when any input is pressed (LED turns off when button is released)