resolve issue 578 (#581)
* resolve issue 578

Fixes issue 578 by adding a check for input mode and applying a bit shift to the mid value present in the 8-bit modes before the shift down to 8-bit. The issue was from the generic gamepad 16-bit mid value being 0x7FFFF shifting down to 0x7F which is not true center for input modes, as they expect 0x80 to be center.

* PS4 macros

Separate PS4 and HID macros, add them to their relevant USB reports.