Use git describe to generate binary version dynamically (#663)
* use the git version in pico version, GP2040VERSION

this uses git to dynamically (at build time) generate the version
string placed into the pico binary metadata (viewable with e.g.
picotool) and in the GP2040VERSION define that is displayed on the
webconfig and parked in the protobuf config

* use the sanitized git version in cmake VERSION

this version string doesn't like as many fancy bits and extra info, but
we can add a bit more context and still have it dynamic this way

* fetch tags in github builds in order to have them available for versioning

* only show the download prompt if currentVersion is not a dev version

'v0.7.5-56-g8f683ae' !== 'v0.7.5', so the webconfig would present the
Get Latest Version button. this change suppresses that, under the
opinion that if you're running anything but a tag, you probably don't
need the download prompt at all. note however that this change also
suppresses the download prompt if you're on v0.7.4-300-gSHA