PS4 latency improvement: force occasional reports (with new counters) in PS4 mode to reduce the need for the PS4 Hack (#792)
* replace the PS4 Hack with occasional forced reports

see comment in the code for the full story, but this fixes some games
(it became a meme for us, but e.g. Ultimate Chicken Horse) that were
broken by the PS4 Hack, with only a minor effect on latency. playing
with the magic may make this tighter still.

* change the PS4 de-Hack to be time-based rather than increment-based

* drop the PS4_KEEPALIVE_TIMER down to 1000 ms

this could probably go down even further, and reduce the pain
experienced by buggy games without affecting latency, but this has
virtually no latency difference compared to the PS4 Hack, so this is a
good starting point and we can find the sweet spot in the future

(...and maybe remove the PS4 Hack don't tell anyone sshhhhhhhh)

* constrain the report counter to 6 bits properly